+++ AV-Solutions/vad-trt/app/autoware_tensorrt_vad/include/autoware/tensorrt_vad/tensor.hpp
